MCS Benchmark
This code reflects work where we eliminated several problems with encoder-based feedback system to the PWM system.

The fault here lies with data handling. Our timing of the if statements in the while loop is far too fast for the actual wheel encoder feedback. For example, the encoder tick variable righttick is 0 for several milliseconds, while the the while(1) loop runs for several microseconds. This results in over-compensation of the PWM manipulation scheme.

Steps forward are to manipulate Timer2 and change the logic (i.e. look at ticks first, then look at time elapsed since last tick).