Tuneable MCS H file
This file is the MCS file with values ready to be tuned. In particular, the values of ks, ki and kpcould be tuned for better performance. With the settings for speed at 5,5, the main program is able to execute a desired path with an inherent bias to the left or right. Current efforts are to tune thse values sothat a straight shot (or close to it) is possible. 
Will vary sampling frequency and speed values as well.